backend/requirements.txt
fastapi
uvicorn
requests
pandas
numpy
networkx
python-dotenv
